requirements.txt
pytest~=7.4.0
selenium~=4.10.0
webdriver-manager
Faker==19.1.0
allure-pytest
requests~=2.31.0
black
isort